+++ src/main/java/com/example/mangoplace/domain/user/controller/TestController.java
@@ -1,4 +1,4 @@
-package com.example.mangoplace.domain.signup.controller;
+package com.example.mangoplace.domain.user.controller;
 
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.CrossOrigin;
