+++ src/main/java/com/example/mangoplace/domain/user/entity/ERole.java
@@ -1,4 +1,4 @@
-package com.example.mangoplace.domain.signup.entity;
+package com.example.mangoplace.domain.user.entity;
 
 public enum ERole {
     ROLE_USER,
