+++ src/main/java/com/example/mangoplace/domain/user/payload/response/JwtResponse.java
@@ -1,4 +1,4 @@
-package com.example.mangoplace.domain.signup.payload.response;
+package com.example.mangoplace.domain.user.payload.response;
 
 import java.util.List;
