+++ src/main/java/com/example/mangoplace/domain/user/security/WebSecurityConfig.java
@@ -1,6 +1,6 @@
-package com.example.mangoplace.domain.signup.security;
+package com.example.mangoplace.domain.user.security;
 
-import com.example.mangoplace.domain.signup.service.UserDetailsServiceImpl;
+import com.example.mangoplace.domain.user.service.UserDetailsServiceImpl;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationManager;
